Rename file to the class name is defined inside
