+++ src/AtlasMapPanel.ts
